add diffitest
