+++ src/main/resources/csrc/wzw_diff.h
@@ -3,4 +3,5 @@
 void wzw_difftest_init(const char *fname);
 int wzw_difftest_diff_and_exec(uint64_t dut_arr[154]);
 void wzw_difftest_set_syn(uint64_t arr[154]);
+extern "C" void inchi_difftest_exec(void);
 #endif
