+++ src/main/scala/system/Configs.scala
@@ -25,7 +25,7 @@ class BaseConfig extends Config(
  * @Description: add the vector config
  */
 // class DefaultConfig extends Config(new WithNBigCores(1) ++ new WithCoherentBusTopology ++ new BaseConfig)
-class DefaultConfig extends Config(new WithVectorUvmNBigCores(1) ++ new WithCoherentBusTopology ++ new BaseConfig)
+class DefaultConfig extends Config(new WithVectorUvmNBigCores(1) ++ new WithCoherentBusTopology ++ new BaseConfig ++ new WithoutEmu)
 
 
 class DefaultBufferlessConfig extends Config(new WithBufferlessBroadcastHub ++ new DefaultConfig)
